RESKC-2579: Checking to see if object is null before deleting. (#2573)

+++ coeus-impl/src/main/java/org/kuali/coeus/propdev/impl/person/ProposalDevelopmentPersonnelController.java
@@ -244,7 +244,9 @@ public class ProposalDevelopmentPersonnelController extends ProposalDevelopmentC
     }
 
     private void deleteCertDetails(ProposalPerson person) {
-        getDataObjectService().delete(person.getCertificationDetails());
+        if (person.getCertificationDetails() != null) {
+            getDataObjectService().delete(person.getCertificationDetails());
+        }
     }
 
     private void deleteProposalPersonBios(DevelopmentProposal proposal, ProposalPerson deleteLine) {
